+++ CHANGELOG.md
@@ -14,6 +14,7 @@ with the following tags indicating the components affected:
 
 ## 2.0.8 - Bugfix release
 - [API - change meaning of ballotsRemainingInCurrentRound][pr88]
+- [INFRA - add dashboard polling load test script][pr89]
 - [UI - Prevent text selection on audit buttons][pr90]
 
 ## 2.0.7 - Bugfix release
@@ -160,4 +161,5 @@ This is [FreeAndFair's most recent tag][1.1.0.3].
 [pr85]: https://github.com/democracyworks/ColoradoRLA/pull/85
 [pr86]: https://github.com/democracyworks/ColoradoRLA/pull/86
 [pr88]: https://github.com/democracyworks/ColoradoRLA/pull/88
+[pr89]: https://github.com/democracyworks/ColoradoRLA/pull/89
 [pr90]: https://github.com/democracyworks/ColoradoRLA/pull/90